RubberDucky
This project provides a security testing tool that performs automated command injection using Rubber Ducky. Rubber Ducky is plugged into a computer as a USB device and is detected as a keyboard, quickly executing commands on the target computer. This project is designed to be used for both security testing and ethical hacking purposes.